Extract base for players
